C2LC-263: Moved to 'fake' strategy for audio manager tests.
