+++ src/FakeAudioManager.js
@@ -1,0 +1,9 @@
+// @flow
+
+import type {AudioManager} from './types';
+
+export default class FakeAudioManager implements AudioManager {
+    playAnnouncement() {};
+    playSoundForCharacterState() {};
+    setAudioEnabled() {};
+};
