+++ src/ProgramBlockEditor.js
@@ -3,13 +3,12 @@
 import { injectIntl, FormattedMessage } from 'react-intl';
 import type {IntlShape} from 'react-intl';
 import * as ProgramUtils from './ProgramUtils';
-import type {Program} from './types';
+import type {AudioManager, Program} from './types';
 import React from 'react';
 import ConfirmDeleteAllModal from './ConfirmDeleteAllModal';
 import AddNode from './AddNode';
 import ActionPanel from './ActionPanel';
 import AriaDisablingButton from './AriaDisablingButton';
-import AudioManager from './AudioManager';
 import FocusTrapManager from './FocusTrapManager';
 import CommandBlock from './CommandBlock';
 import classNames from 'classnames';
